Reorganize matrices
